+++ app/src/main/java/eu/kanade/domain/DomainModule.kt
@@ -49,23 +49,22 @@ import eu.kanade.domain.track.manga.interactor.RefreshMangaTracks
 import eu.kanade.domain.track.manga.interactor.SyncChapterProgressWithTrack
 import eu.kanade.domain.track.manga.interactor.TrackChapter
 import eu.kanade.tachiyomi.ui.player.utils.TrackSelect
-import mihon.data.repository.anime.AnimeExtensionRepoRepositoryImpl
-import mihon.data.repository.manga.MangaExtensionRepoRepositoryImpl
-import mihon.domain.extensionrepo.anime.interactor.CreateAnimeExtensionRepo
-import mihon.domain.extensionrepo.anime.interactor.DeleteAnimeExtensionRepo
-import mihon.domain.extensionrepo.anime.interactor.GetAnimeExtensionRepo
-import mihon.domain.extensionrepo.anime.interactor.GetAnimeExtensionRepoCount
-import mihon.domain.extensionrepo.anime.interactor.ReplaceAnimeExtensionRepo
-import mihon.domain.extensionrepo.anime.interactor.UpdateAnimeExtensionRepo
-import mihon.domain.extensionrepo.anime.repository.AnimeExtensionRepoRepository
-import mihon.domain.extensionrepo.manga.interactor.CreateMangaExtensionRepo
-import mihon.domain.extensionrepo.manga.interactor.DeleteMangaExtensionRepo
-import mihon.domain.extensionrepo.manga.interactor.GetMangaExtensionRepo
-import mihon.domain.extensionrepo.manga.interactor.GetMangaExtensionRepoCount
-import mihon.domain.extensionrepo.manga.interactor.ReplaceMangaExtensionRepo
-import mihon.domain.extensionrepo.manga.interactor.UpdateMangaExtensionRepo
-import mihon.domain.extensionrepo.manga.repository.MangaExtensionRepoRepository
-import mihon.domain.extensionrepo.service.ExtensionRepoService
+import mihon.data.extension.repository.AnimeExtensionStoreRepositoryImpl
+import mihon.data.extension.repository.MangaExtensionStoreRepositoryImpl
+import mihon.data.extension.service.AnimeExtensionStoreService
+import mihon.data.extension.service.MangaExtensionStoreService
+import mihon.domain.extension.anime.interactor.AddAnimeExtensionStore
+import mihon.domain.extension.anime.interactor.GetAnimeExtensionStoreCountAsFlow
+import mihon.domain.extension.anime.interactor.GetAnimeExtensionStores
+import mihon.domain.extension.anime.interactor.RemoveAnimeExtensionStore
+import mihon.domain.extension.anime.interactor.UpdateAnimeExtensionStores
+import mihon.domain.extension.anime.repository.AnimeExtensionStoreRepository
+import mihon.domain.extension.manga.interactor.AddMangaExtensionStore
+import mihon.domain.extension.manga.interactor.GetMangaExtensionStoreCountAsFlow
+import mihon.domain.extension.manga.interactor.GetMangaExtensionStores
+import mihon.domain.extension.manga.interactor.RemoveMangaExtensionStore
+import mihon.domain.extension.manga.interactor.UpdateMangaExtensionStores
+import mihon.domain.extension.manga.repository.MangaExtensionStoreRepository
 import mihon.domain.items.chapter.interactor.FilterChaptersForDownload
 import mihon.domain.items.episode.interactor.FilterEpisodesForDownload
 import mihon.domain.upcoming.anime.interactor.GetUpcomingAnime
@@ -399,25 +398,24 @@ class DomainModule : InjektModule {
         addFactory { TrustAnimeExtension(get(), get()) }
         addFactory { TrustMangaExtension(get(), get()) }
 
-        addFactory { ExtensionRepoService(get(), get()) }
+        addFactory { AnimeExtensionStoreService(get(), get(), get()) }
+        addFactory { MangaExtensionStoreService(get(), get(), get()) }
 
-        addSingletonFactory<AnimeExtensionRepoRepository> { AnimeExtensionRepoRepositoryImpl(get()) }
-        addFactory { GetAnimeExtensionRepo(get()) }
-        addFactory { GetAnimeExtensionRepoCount(get()) }
-        addFactory { CreateAnimeExtensionRepo(get(), get()) }
-        addFactory { DeleteAnimeExtensionRepo(get()) }
-        addFactory { ReplaceAnimeExtensionRepo(get()) }
-        addFactory { UpdateAnimeExtensionRepo(get(), get()) }
+        addSingletonFactory<AnimeExtensionStoreRepository> { AnimeExtensionStoreRepositoryImpl(get(), get()) }
+        addFactory { GetAnimeExtensionStores(get()) }
+        addFactory { GetAnimeExtensionStoreCountAsFlow(get()) }
+        addFactory { AddAnimeExtensionStore(get()) }
+        addFactory { RemoveAnimeExtensionStore(get()) }
+        addFactory { UpdateAnimeExtensionStores(get()) }
         addFactory { ToggleAnimeIncognito(get()) }
         addFactory { GetAnimeIncognitoState(get(), get(), get()) }
 
-        addSingletonFactory<MangaExtensionRepoRepository> { MangaExtensionRepoRepositoryImpl(get()) }
-        addFactory { GetMangaExtensionRepo(get()) }
-        addFactory { GetMangaExtensionRepoCount(get()) }
-        addFactory { CreateMangaExtensionRepo(get(), get()) }
-        addFactory { DeleteMangaExtensionRepo(get()) }
-        addFactory { ReplaceMangaExtensionRepo(get()) }
-        addFactory { UpdateMangaExtensionRepo(get(), get()) }
+        addSingletonFactory<MangaExtensionStoreRepository> { MangaExtensionStoreRepositoryImpl(get(), get()) }
+        addFactory { GetMangaExtensionStores(get()) }
+        addFactory { GetMangaExtensionStoreCountAsFlow(get()) }
+        addFactory { AddMangaExtensionStore(get()) }
+        addFactory { RemoveMangaExtensionStore(get()) }
+        addFactory { UpdateMangaExtensionStores(get()) }
         addFactory { ToggleMangaIncognito(get()) }
         addFactory { GetMangaIncognitoState(get(), get(), get()) }