+++ app/src/main/java/eu/kanade/tachiyomi/data/backup/restore/restorers/MangaRestorer.kt
@@ -280,6 +280,7 @@ class MangaRestorer(
                 notes = manga.notes,
             )
         }
+    }
 
     private suspend fun restoreMangaDetails(
         manga: Manga,