+++ app/src/main/java/eu/kanade/tachiyomi/extension/anime/api/AnimeExtensionApi.kt
@@ -1,83 +1,34 @@
 package eu.kanade.tachiyomi.extension.anime.api
 
 import android.content.Context
-import eu.kanade.domain.source.service.SourcePreferences
 import eu.kanade.tachiyomi.extension.ExtensionUpdateNotifier
 import eu.kanade.tachiyomi.extension.anime.AnimeExtensionManager
 import eu.kanade.tachiyomi.extension.anime.model.AnimeExtension
 import eu.kanade.tachiyomi.extension.anime.model.AnimeLoadResult
 import eu.kanade.tachiyomi.extension.anime.util.AnimeExtensionLoader
-import eu.kanade.tachiyomi.network.GET
-import eu.kanade.tachiyomi.network.NetworkHelper
-import eu.kanade.tachiyomi.network.awaitSuccess
-import eu.kanade.tachiyomi.network.parseAs
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import logcat.LogPriority
-import mihon.domain.extensionrepo.anime.interactor.GetAnimeExtensionRepo
-import mihon.domain.extensionrepo.anime.interactor.UpdateAnimeExtensionRepo
-import mihon.domain.extensionrepo.model.ExtensionRepo
+import mihon.domain.extension.anime.interactor.UpdateAnimeExtensionStores
+import mihon.domain.extension.anime.repository.AnimeExtensionStoreRepository
 import tachiyomi.core.common.preference.Preference
 import tachiyomi.core.common.preference.PreferenceStore
 import tachiyomi.core.common.util.lang.withIOContext
-import tachiyomi.core.common.util.system.logcat
 import uy.kohesive.injekt.injectLazy
 import java.time.Instant
 import kotlin.time.Duration.Companion.days
 
 internal class AnimeExtensionApi {
 
-    private val networkService: NetworkHelper by injectLazy()
+    private val repository: AnimeExtensionStoreRepository by injectLazy()
     private val preferenceStore: PreferenceStore by injectLazy()
-    private val getExtensionRepo: GetAnimeExtensionRepo by injectLazy()
-    private val updateExtensionRepo: UpdateAnimeExtensionRepo by injectLazy()
-    private val animeExtensionManager: AnimeExtensionManager by injectLazy()
-    private val json: Json by injectLazy()
-    private val sourcePreferences: SourcePreferences by injectLazy()
+    private val updateExtensionStores: UpdateAnimeExtensionStores by injectLazy()
+    private val extensionManager: AnimeExtensionManager by injectLazy()
 
     private val lastExtCheck: Preference<Long> by lazy {
         preferenceStore.getLong("last_ext_check", 0)
     }
 
+    @Suppress("UNCHECKED_CAST")
     suspend fun findExtensions(): List<AnimeExtension.Available> {
-        // KMK -->
-        val disabledRepos = sourcePreferences.disabledRepos.get()
-        // KMK <--
-        return withIOContext {
-            getExtensionRepo.getAll()
-                // KMK -->
-                .filterNot { it.baseUrl in disabledRepos }
-                // KMK <--
-                .map { async { getExtensions(it) } }
-                .awaitAll()
-                .flatten()
-        }
-    }
-
-    private suspend fun getExtensions(extRepo: ExtensionRepo): List<AnimeExtension.Available> {
-        val repoBaseUrl = extRepo.baseUrl
-        return try {
-            val response = networkService.client
-                .newCall(GET("$repoBaseUrl/index.min.json"))
-                .awaitSuccess()
-
-            with(json) {
-                response
-                    .parseAs<List<AnimeExtensionJsonObject>>()
-                    .toExtensions(
-                        repoBaseUrl,
-                        // KMK -->
-                        signature = extRepo.signingKeyFingerprint,
-                        repoName = extRepo.shortName ?: extRepo.name,
-                        // KMK <--
-                    )
-            }
-        } catch (e: Throwable) {
-            logcat(LogPriority.ERROR, e) { "Failed to get extensions from $repoBaseUrl" }
-            emptyList()
-        }
+        return withIOContext { repository.fetchExtensions() as List<AnimeExtension.Available> }
     }
 
     suspend fun checkForUpdates(
@@ -91,11 +42,10 @@ internal class AnimeExtensionApi {
             return null
         }
 
-        // Update extension repo details
-        updateExtensionRepo.awaitAll()
+        updateExtensionStores()
 
         val extensions = if (fromAvailableExtensionList) {
-            animeExtensionManager.availableExtensionsFlow.value
+            extensionManager.availableExtensionsFlow.value
         } else {
             findExtensions().also { lastExtCheck.set(Instant.now().toEpochMilli()) }
         }
@@ -108,7 +58,6 @@ internal class AnimeExtensionApi {
         for (installedExt in installedExtensions) {
             val pkgName = installedExt.pkgName
             val availableExt = extensions.find { it.pkgName == pkgName } ?: continue
-
             val hasUpdatedVer = availableExt.versionCode > installedExt.versionCode
             val hasUpdatedLib = availableExt.libVersion > installedExt.libVersion
             val hasUpdate = hasUpdatedVer || hasUpdatedLib
@@ -126,76 +75,4 @@ internal class AnimeExtensionApi {
 
         return extensionsWithUpdate
     }
-
-    private fun List<AnimeExtensionJsonObject>.toExtensions(
-        repoUrl: String,
-        // KMK -->
-        signature: String,
-        repoName: String,
-        // KMK <--
-    ): List<AnimeExtension.Available> {
-        return this
-            .filter {
-                val libVersion = it.extractLibVersion()
-                libVersion >= AnimeExtensionLoader.LIB_VERSION_MIN && libVersion <= AnimeExtensionLoader.LIB_VERSION_MAX
-            }
-            .map {
-                AnimeExtension.Available(
-                    name = it.name.substringAfter("Animetail: "),
-                    pkgName = it.pkg,
-                    versionName = it.version,
-                    versionCode = it.code,
-                    libVersion = it.extractLibVersion(),
-                    lang = it.lang,
-                    isNsfw = it.nsfw == 1,
-                    isTorrent = it.torrent == 1,
-                    sources = it.sources?.map(extensionAnimeSourceMapper).orEmpty(),
-                    apkName = it.apk,
-                    iconUrl = "$repoUrl/icon/${it.pkg}.png",
-                    repoUrl = repoUrl,
-                    // KMK -->
-                    signatureHash = signature,
-                    repoName = repoName,
-                    // KMK <--
-                )
-            }
-    }
-
-    fun getApkUrl(extension: AnimeExtension.Available): String {
-        return "${extension.repoUrl}/apk/${extension.apkName}"
-    }
-
-    private fun AnimeExtensionJsonObject.extractLibVersion(): Double {
-        return version.substringBeforeLast('.').toDouble()
-    }
-}
-
-@Serializable
-private data class AnimeExtensionJsonObject(
-    val name: String,
-    val pkg: String,
-    val apk: String,
-    val lang: String,
-    val code: Long,
-    val version: String,
-    val nsfw: Int,
-    val torrent: Int = 0,
-    val sources: List<AnimeExtensionSourceJsonObject>?,
-)
-
-@Serializable
-private data class AnimeExtensionSourceJsonObject(
-    val id: Long,
-    val lang: String,
-    val name: String,
-    val baseUrl: String,
-)
-
-private val extensionAnimeSourceMapper: (AnimeExtensionSourceJsonObject) -> AnimeExtension.Available.AnimeSource = {
-    AnimeExtension.Available.AnimeSource(
-        id = it.id,
-        lang = it.lang,
-        name = it.name,
-        baseUrl = it.baseUrl,
-    )
 }